Moved getSimpleNameOfAnObjectsClass, parseSimpleName and stripDollars to NameUtil.
