+++ scalactic-macro/src/main/scala/org/scalactic/NameUtil.scala
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2001-2018 Artima, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.scalactic
+
+private[org] object NameUtil {
+
+  // This attempts to strip dollar signs that happen when using the interpreter. It is quite fragile
+  // and already broke once. In the early days, all funky dollar sign encrusted names coming out of
+  // the interpreter started with "line". Now they don't, but in both cases they seemed to have at
+  // least one "$iw$" in them. So now I leave the string alone unless I see a "$iw$" in it. Worst case
+  // is sometimes people will get ugly strings coming out of the interpreter. -bv April 3, 2012
+  def stripDollars(s: String): String = {
+    val lastDollarIndex = s.lastIndexOf('$')
+    if (lastDollarIndex < s.length - 1)
+      if (lastDollarIndex == -1 || !s.contains("$iw$")) s else s.substring(lastDollarIndex + 1)
+    else {
+      // The last char is a dollar sign
+      val lastNonDollarChar = s.reverse.find(_ != '$')
+      lastNonDollarChar match {
+        case None => s
+        case Some(c) => {
+          val lastNonDollarIndex = s.lastIndexOf(c)
+          if (lastNonDollarIndex == -1) s
+          else stripDollars(s.substring(0, lastNonDollarIndex + 1))
+        }
+      }
+    }
+  }
+
+  def getSimpleNameOfAnObjectsClass(o: AnyRef) = stripDollars(parseSimpleName(o.getClass.getName))
+
+  // [bv: this is a good example of the expression type refactor. I moved this from SuiteClassNameListCellRenderer]
+  // this will be needed by the GUI classes, etc.
+  def parseSimpleName(fullyQualifiedName: String) = {
+
+    val dotPos = fullyQualifiedName.lastIndexOf('.')
+
+    // [bv: need to check the dotPos != fullyQualifiedName.length]
+    if (dotPos != -1 && dotPos != fullyQualifiedName.length)
+      fullyQualifiedName.substring(dotPos + 1)
+    else
+      fullyQualifiedName
+  }
+
+}
